pyproject.toml
[tool.poetry]
name = "summit_sphere"
version = "0.1.0"
description = ""
authors = ["Piotr Pawełczak <pawelczakpiotr99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-extensions = "^3.2.3"
pytest = "^7.4.3"
ruff = "^0.1.6"
mypy = "^1.7.1"
pre-commit = "^3.5.0"
psycopg2 = "^2.9.9"
pytest-django = "^4.7.0"
bpython = "^0.24"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^4.2.0"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
